pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-character-studio-ui"
version = "0.1.0"
description = "Constitution-driven persona training studio built on the Tinker SDK."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31",
  "datasets>=2.19",
  "transformers>=4.43",
  "streamlit>=1.37",
  "pydantic>=2.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
tinker = ["tinker>=0.3.0"]
train = [
  "torch",  # install matching CUDA/CPU build for your platform
  "tinker>=0.3.0",
  "transformers>=4.43",
  "datasets>=2.19",
]
studio = [
  "streamlit>=1.37",
]
dev = [
  "pytest>=8.3",
  "ruff",
  "black",
  "isort",
]
all = [
  "tinker>=0.3.0",
  "torch",
  "transformers>=4.43",
  "datasets>=2.19",
  "streamlit>=1.37",
  "pytest>=8.3",
  "ruff",
  "black",
  "isort",
]

[tool.setuptools]
packages = ["character", "studio"]
